IMPROVE: Make this script work with Ubuntu variants.
There are many Ubuntu variants such as Linux Mint which could use this
script.

For this approach, I've changed the checking from `lsb_release` to /etc/os-release
since it has more enviroment variables we could use and make this script
portable to other Linux variants in the future.

This script was tested with LinuxMint 19 - Tara, which uses Ubuntu 18.04 packages.